requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
anthropic==0.34.2
google-api-python-client==2.143.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
chromadb==0.5.15
sentence-transformers==3.1.1
PyMuPDF==1.24.11
fpdf2==2.7.9
Pillow==10.4.0
aiofiles==24.1.0
sqlalchemy==2.0.35
pydantic-settings==2.5.2
python-multipart==0.0.12
aioredis==2.0.1
pytest==8.0.0
pytest-asyncio==0.23.0
pytest-mock==3.14.0
